refactor(frontend): refine holiday domain implementation
Renamed the "domains" folder as "features"
Removed unused variables
