+++ resource-app/frontend/features/holiday/types.ts
