+++ resource-app/frontend/hooks/useCalendar.ts
@@ -1,7 +1,7 @@
 
 import { useState, useMemo, useCallback } from 'react';
 import { useApp } from '../context/AppContext';
-import { useHoliday } from '../domains/holiday/context';
+import { useHoliday } from '../features/holiday/context';
 import { BookingStatus, UserRole } from '../types';
 import { endOfMonth, eachDayOfInterval, isSameDay, getDay, format } from 'date-fns';
